pyproject.toml
[build-system]
requires = ["hatchling", "grpcio-tools>=1.50.0"]
build-backend = "hatchling.build"

[project]
name = "salesforce-zerobus"
version = "1.0.0"
description = "Simple Python library for streaming Salesforce Change Data Capture events to Databricks Delta tables"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Grant Doyle", email = "grant.doyle@databricks.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
keywords = ["salesforce", "databricks", "streaming", "cdc", "change-data-capture", "real-time"]
requires-python = ">=3.10"

dependencies = [
    # Core dependencies
    "grpcio>=1.50.0",
    "grpcio-tools>=1.50.0", 
    "protobuf>=4.0.0,<7.0.0",
    "avro-python3>=1.10.0",
    
    # Databricks dependencies
    "databricks-sdk>=0.12.0",
    "databricks-labs-lsql>=0.2.0",
    
    # Utility dependencies
    "bitstring>=4.0.0",
    "certifi>=2022.0.0", 
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
    
    # Note: databricks-zerobus will be bundled directly in the wheel
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0", 
    "isort>=5.10.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/DoyleDev/Salesforce-Zerobus"
Documentation = "https://github.com/DoyleDev/Salesforce-Zerobus/blob/main/README.md"
Repository = "https://github.com/DoyleDev/Salesforce-Zerobus.git"
Issues = "https://github.com/DoyleDev/Salesforce-Zerobus/issues"

# Hatchling build configuration
[tool.hatch.build]
include = [
    "salesforce_zerobus/",
    "wheels/databricks_zerobus-0.0.16-py3-none-any.whl",
]

[tool.hatch.build.targets.wheel]
packages = ["salesforce_zerobus"]

# Allow direct references to local files
[tool.hatch.metadata]
allow-direct-references = true

# Development tools configuration  
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # Exclude generated protobuf files
  .*_pb2\.py
  | .*_pb2_grpc\.py
)/
'''

[tool.isort]  
profile = "black"
line_length = 100
skip_glob = ["*_pb2.py", "*_pb2_grpc.py"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    ".*_pb2.py",
    ".*_pb2_grpc.py", 
]